Moved component input from styles
